Many Features implemented

+++ model/gameModel.js
@@ -5,7 +5,6 @@ let getYourSanta = (G_ID, U_ID) =>
 		const sql = 'SELECT * FROM `user_game`AS ug LEFT JOIN `user` AS u ON ug.U_ID = u.U_ID WHERE ug.recipient = ? AND ug.G_ID = ?;';
 		db.query(sql, [U_ID, G_ID], (error, results) => {
 			if (error) reject(error);
-			console.log('Santa: ', results);
 			resolve(results);
 		});
 	});
@@ -15,7 +14,6 @@ let getYourTarget = (G_ID, U_ID) =>
 		const sql = 'SELECT * FROM `user_game`AS ug LEFT JOIN `user` AS u ON ug.recipient = u.U_ID WHERE ug.U_ID = ? AND ug.G_ID = ?;';
 		db.query(sql, [U_ID, G_ID], (error, results) => {
 			if (error) reject(error);
-			console.log('Target: ', results);
 			resolve(results);
 		});
 	});
@@ -23,7 +21,6 @@ let getYourTarget = (G_ID, U_ID) =>
 function create({ game_name, description, member_count, number_of_circles, visibility }, creator) {
 	return new Promise((resolve, reject) => {
 		const request = 'INSERT INTO `game`(`name`, `description`, `creator`, `max_members` ,`visibility`) VALUES (?, ?, ?, ?,?)';
-		console.log(request, [game_name, description, creator, member_count, visibility]);
 		db.query(request, [game_name, description, creator, member_count, visibility], (err, res) => {
 			if (err) reject(err);
 			if (res && res.affectedRows > 0) resolve({ G_ID: res.insertId });
@@ -52,7 +49,7 @@ let getMembers = (gameId) =>
 
 let getGames = (U_ID) =>
 	new Promise((resolve, reject) => {
-		const sql = 'SELECT * FROM `game` WHERE `creator` = ' + db.escape(U_ID);
+		const sql = 'SELECT * FROM `game` AS g LEFT JOIN user_game AS ug ON g.G_ID = ug.G_ID  WHERE `creator` = ' + db.escape(U_ID) + ' OR `U_ID` = ' + db.escape(U_ID) + ' GROUP BY g.G_ID';
 		db.query(sql, (error, results) => {
 			if (error) reject(error);
 			resolve(results);
@@ -102,10 +99,8 @@ let assignSantas = (assignedMembers, G_ID) =>
 
 		const params = updateValues.concat(G_ID);
 
-		console.log(query, [G_ID]);
 		db.query(query, [G_ID], (error, results) => {
 			if (error) reject(error);
-			console.log('Santas assigned: ', results);
 			resolve(results);
 		});
 	});
@@ -124,20 +119,29 @@ function update(gameId, updatedData) {
 		const request = 'UPDATE `game` SET ? WHERE `G_ID` = ?';
 		db.query(request, [updatedData, gameId], (err, res) => {
 			if (err) reject(err);
-			console.log(res);
 			if (res && res.affectedRows > 0) resolve({ message: 'Game updated successfully' });
 			else reject(new Error('Game not found'));
 		});
 	});
 }
 
-function remove(gameId) {
+function deleteGame(gameId, U_ID) {
 	return new Promise((resolve, reject) => {
-		const request = 'DELETE FROM `game` WHERE `G_ID` = ?';
-		db.query(request, [gameId], (err, res) => {
+		const request = 'DELETE FROM `game` WHERE `G_ID` = ? AND `creator` = ?';
+		db.query(request, [gameId, U_ID], (err, res) => {
 			if (err) reject(err);
-			console.log(res);
 			if (res && res.affectedRows > 0) resolve({ message: 'Game deleted successfully' });
+			else reject(new Error('Game not found'));
+		});
+	});
+}
+
+function remove(gameId, userId) {
+	return new Promise((resolve, reject) => {
+		const request = 'DELETE FROM `user_game` WHERE `G_ID` = ? AND `U_ID` = ?';
+		db.query(request, [gameId, userId], (err, res) => {
+			if (err) reject(err);
+			if (res && res.affectedRows > 0) resolve({ message: 'User removed' });
 			else reject(new Error('Game not found'));
 		});
 	});
@@ -148,7 +152,7 @@ let checkAdmin = (G_ID, U_ID) =>
 		const sql = 'SELECT * FROM `game` WHERE `G_ID` = ? AND `creator` = ?';
 		db.query(sql, [G_ID, U_ID], (error, results) => {
 			if (error) reject(error);
-			resolve(results.length >= 0);
+			resolve(results.length > 0);
 		});
 	});
 
@@ -166,7 +170,6 @@ let deleteAllChats = (G_ID) =>
 		const sql = 'DELETE FROM `chat` WHERE `game` = ' + G_ID;
 		db.query(sql, (error, results) => {
 			if (error) reject(error);
-			console.log('Chats deleted: ', results);
 			resolve(results);
 		});
 	});
@@ -183,18 +186,74 @@ let createChats = (members, G_ID) =>
 		const sql = sqlStart + sqlMiddle + sqlEnd;
 		db.query(sql, (error, results) => {
 			if (error) reject(error);
-			console.log('Chats created: ', results);
 			resolve(results);
 		});
 	});
 
 let chatSanta = (G_ID, U_ID) =>
 	new Promise((resolve, reject) => {
-		const sql = 'SELECT * FROM message AS m LEFT JOIN chat AS c ON m.chat = c.C_ID WHERE c.game = ? AND c.user_receiver = ?';
+		const sql = 'SELECT * FROM message AS m LEFT JOIN chat AS c ON m.chat = c.C_ID WHERE c.game = ? AND c.user_receiver = ? ORDER BY m.timestamp DESC';
 		db.query(sql, [G_ID, U_ID], (error, results) => {
 			if (error) reject(error);
-			console.log('Chat created: ', results);
 			resolve(results);
+		});
+	});
+
+let chatSantaId = (G_ID, U_ID) =>
+	new Promise((resolve, reject) => {
+		const sql = 'SELECT C_ID FROM chat WHERE user_receiver = ? AND game = ?';
+		db.query(sql, [U_ID, G_ID], (error, results) => {
+			if (error) reject(error);
+			resolve(results[0].C_ID);
+		});
+	});
+
+let sendMessage = (chatId, sender, message) =>
+	new Promise((resolve, reject) => {
+		const sql = 'INSERT INTO `message`(`chat`, `message_content`, `sender`) VALUES (?, ?, ?)';
+		db.query(sql, [chatId, message, sender], (error, results) => {
+			if (error) reject(error);
+			resolve(results);
+		});
+	});
+
+let chatRecipient = (G_ID, U_ID) =>
+	new Promise((resolve, reject) => {
+		const sql = 'SELECT * FROM message AS m LEFT JOIN chat AS c ON m.chat = c.C_ID WHERE c.game = ? AND c.user_giver = ? ORDER BY m.timestamp DESC';
+		db.query(sql, [G_ID, U_ID], (error, results) => {
+			if (error) reject(error);
+			resolve(results);
+		});
+	});
+
+let chatRecipientId = (G_ID, U_ID) =>
+	new Promise((resolve, reject) => {
+		const sql = 'SELECT C_ID FROM chat WHERE user_giver = ? AND game = ?';
+		db.query(sql, [U_ID, G_ID], (error, results) => {
+			if (error) reject(error);
+			if (results.length > 0) {
+				resolve(results[0].C_ID);
+			} else reject('No chat found');
+		});
+	});
+let getRecipient = (G_ID, U_ID) =>
+	new Promise((resolve, reject) => {
+		const sql = 'SELECT * FROM chat AS c LEFT JOIN `user` AS u ON c.user_receiver = u.U_ID WHERE c.user_giver = ? AND c.game = ? GROUP BY c.user_receiver';
+		db.query(sql, [U_ID, G_ID], (error, results) => {
+			if (error) reject(error);
+			if (results.length > 0) {
+				resolve(results[0]);
+			} else reject('No chat found');
+		});
+	});
+let getSanta = (G_ID, U_ID) =>
+	new Promise((resolve, reject) => {
+		const sql = 'SELECT * FROM chat AS c LEFT JOIN `user` AS u ON c.user_giver = u.U_ID WHERE c.user_receiver = ? AND c.game = ? GROUP BY c.user_giver';
+		db.query(sql, [U_ID, G_ID], (error, results) => {
+			if (error) reject(error);
+			if (results.length > 0) {
+				resolve(results[0]);
+			} else reject('No chat found');
 		});
 	});
 
@@ -202,7 +261,8 @@ module.exports = {
 	create,
 	get,
 	update,
-	delete: remove,
+	remove,
+	deleteGame,
 	getGames,
 	getMembers,
 	joinGame,
@@ -216,4 +276,11 @@ module.exports = {
 	checkAdmin,
 	deleteAllChats,
 	createChats,
+	chatSanta,
+	chatSantaId,
+	sendMessage,
+	chatRecipient,
+	chatRecipientId,
+	getRecipient,
+	getSanta,
 };
